+++ benchmark_analyzer/Cargo.toml
@@ -1,10 +1,9 @@
 [package]
-name = "benchmark-analyzer"
+name = "benchmark_analyzer"
 version = "0.1.0"
 edition = "2021"
 
 [lib]
 name = "benchmark"
-path = "src/tools.rs"  
 
 [dependencies]
