+++ crates/abi/Cargo.toml
@@ -6,3 +6,4 @@ edition = "2021"
 [dependencies]
 miralis_core = { path = "../core" }
 log = { workspace = true }
+config_helpers = { path = "../config_helpers"}
