+++ firmware/benchmark/csr_write/Cargo.toml
@@ -1,0 +1,11 @@
+[package]
+name = "csr_write"
+version = "0.1.0"
+edition = "2021"
+
+[[bin]]
+name = "csr_write"
+path = "main.rs"
+
+[dependencies]
+miralis_abi = { path = "../../../crates/abi" }
