--- firmware/csr_write_benchmark/Cargo.toml
@@ -1,11 +1,0 @@
-[package]
-name = "csr_write_benchmark"
-version = "0.1.0"
-edition = "2021"
-
-[[bin]]
-name = "csr_write_benchmark"
-path = "main.rs"
-
-[dependencies]
-miralis_abi = { path = "../../crates/abi" }
