+++ runner/Cargo.toml
@@ -7,4 +7,5 @@ edition = "2021"
 clap = { version = "4.4.11", features = ["derive"] }
 serde = { version = "1.0.196", features = ["derive"] }
 toml = "0.8.10"
-benchmark-analyzer = { path = "../benchmark-analyzer" }
+benchmark_analyzer = { path = "../benchmark_analyzer" }
+config_helpers = { path = "../crates/config_helpers" }
